refactor(ka): move tools/custom from pkg/ to internal/ (#995)
tools/custom depends on internal/kubernautagent/enrichment interfaces,
so it belongs in internal/, not pkg/. The only production consumer is
cmd/kubernautagent/main.go. Fixes the pkg->internal layering violation
by placing the code where it belongs rather than promoting internal
types to pkg/.

Updates import paths in 11 consumer files.
