+++ internal/kubernautagent/tools/custom/resource_context.go
