+++ test/unit/kubernautagent/tools/custom/context_cancel_779_test.go
@@ -24,7 +24,7 @@ import (
 	. "github.com/onsi/gomega"
 
 	katypes "github.com/jordigilh/kubernaut/pkg/kubernautagent/types"
-	"github.com/jordigilh/kubernaut/pkg/kubernautagent/tools/custom"
+	"github.com/jordigilh/kubernaut/internal/kubernautagent/tools/custom"
 )
 
 // BR-WORKFLOW-016 / #779: Tool Execute methods must not panic when the
