+++ test/unit/kubernautagent/tools/custom/custom_tools_test.go
@@ -29,7 +29,7 @@ import (
 
 	katypes "github.com/jordigilh/kubernaut/pkg/kubernautagent/types"
 	ogenclient "github.com/jordigilh/kubernaut/pkg/datastorage/ogen-client"
-	"github.com/jordigilh/kubernaut/pkg/kubernautagent/tools/custom"
+	"github.com/jordigilh/kubernaut/internal/kubernautagent/tools/custom"
 )
 
 // toolCtx returns a context with a representative SignalContext attached.
