+++ test/unit/kubernautagent/tools/custom/ds_error_paths_779_test.go
@@ -28,7 +28,7 @@ import (
 
 	katypes "github.com/jordigilh/kubernaut/pkg/kubernautagent/types"
 	ogenclient "github.com/jordigilh/kubernaut/pkg/datastorage/ogen-client"
-	"github.com/jordigilh/kubernaut/pkg/kubernautagent/tools/custom"
+	"github.com/jordigilh/kubernaut/internal/kubernautagent/tools/custom"
 )
 
 // errorDS returns errors from all DS methods.
